1.Ble coderefactor. 2.Trace file code Refactor. 3.Rxdata parser code refactor
